% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBI.R
\name{LBI}
\alias{LBI}
\title{LBI}
\usage{
LBI(
  study_area_id,
  study_area_polygon,
  nobs = 250,
  nboot = 1000,
  fact_location = 10,
  elevation = NULL,
  bootstrap_output = FALSE
)
}
\arguments{
\item{study_area_id}{Character string. Name ID of the study or country.}

\item{study_area_polygon}{Polygon shape}

\item{nobs}{Numeric, number of random observations in each sample
250 by default}

\item{nboot}{Numeric, null model iterations. 1,000 by default}

\item{fact_location}{Numeric. All possible coordinate locations that can be
sampled (value that multiplies nobs to predefined the coordinate locations
that can be included in each iteration; aims to optimize the calculation
time).}

\item{elevation}{Elevation raster. elevation in wgs84; if not provided,
NA will be returned for null-model elevational shifts.}

\item{bootstrap_output}{Logical. FALSE by default. If TRUE, all bootstraps
are returned as a data.frame.}
}
\value{
A data.frame or a list of two data.frames if bootstrap_output is set to TRUE.
}
\description{
Computes the Latitudinal Bias Index (LBI) for a given shapefile, by
calculating the distance between two random locations within the shape
multiple times (see details).
}
\details{
The main output contains the following columns:
\itemize{
\item study_area_id: ID or name of the study case region
\item abs_distance_km: average expected distance shift between t1 and t2
\item null_mod_NS_shift: average expected North-South shift between t1 and t2
\item null_mod_EW_shift: average expected East-West shift between t1 and t2
\item null_mod_elevation_shift: average expected elevation shift between t1 and
t2
\item LBI: the Latitudinal Bias Index value
}
}
\examples{

study_area <- rnaturalearth::ne_countries(scale = 110, continent = "Europe",
country = "Sweden", type = "map_units", returnclass = "sf")
study_area <- sf::st_union(study_area)

LBI(study_area_id = "Sweden", study_area_polygon = study_area,
nobs = 10, nboot = 10, fact_location = 10, elevation = NULL)
\donttest{
# With elevation
elevation_df <- elevatr::get_elev_raster(
locations = sf::st_as_sf(study_area), z = 5)
}

}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Sanczuk et al. submitted.
}\if{html}{\out{</div>}}
}
