% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{windrose}
\alias{windrose}
\title{windrose}
\usage{
windrose(
  data,
  spd,
  dir,
  spdres = 30,
  dirres = 30,
  spdmin = 0,
  spdmax = 150,
  spdseq = NULL,
  palette = "Spectral",
  countmax = NA,
  debug = 0
)
}
\arguments{
\item{data}{data frame with bearing and speed}

\item{spd}{speed column}

\item{dir}{bearing column}

\item{spdres}{= 30, #adapt here the legend and breaks}

\item{dirres}{= 30,  #resolution of bearings}

\item{spdmin}{= 0, #and here min}

\item{spdmax}{= 150,#and here max; check as max}

\item{spdseq}{= NULL}

\item{palette}{= "Spectral"}

\item{countmax}{= NA}

\item{debug}{=}
}
\value{
A plot.
}
\description{
Plot the directional shifts for a given bootstrap.
}
\details{
Windrose.
}
\examples{

\donttest{
study_area <- rnaturalearth::ne_countries(scale = 110, continent = "Europe",
country = "Sweden", type = "map_units", returnclass = "sf")
study_area <- sf::st_union(study_area)
po <- LBI(study_area_id = "Sweden", study_area_polygon = study_area,
nobs = 10, nboot = 10, fact_location = 10, elevation = NA,
bootstrap_output = TRUE)
test <- as.data.frame(po$all[, c("study_area_id", "abs_distance_km",
"bearing", "rep")])
test$abs_distance_km <- as.numeric(test$abs_distance_km)
pop <- windrose(data = test, spd = "abs_distance_km", dir = "bearing")
}

}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Sanczuk et al. submitted.
}\if{html}{\out{</div>}}
}
